Update README and state_variables for cadCAD migration
