--- cadcad/engine/__init__.py
@@ -1,2 +1,0 @@
-"""empty (and it should be for clean namespaces)
-"""
