+++ cadcad/simulation_abm/model/SimState.py
@@ -1,0 +1,119 @@
+import logging
+log = logging.getLogger('simstate')
+
+from enforce_typing import enforce_types # type: ignore[import]
+from typing import Set
+
+from .parts.agents.GrantTakingAgent import GrantTakingAgent
+from .SimStrategy import SimStrategy
+from .parts.stats.Kpis import KPIs
+from .parts.util import mathutil, valuation
+from .parts.util.mathutil import Range
+from .parts.util.constants import *
+
+@enforce_types
+class SimState(object):
+    
+    def __init__(self, ss: SimStrategy):
+        log.debug("init:begin")
+        
+        #main
+        self.ss = ss
+        self.tick = 0
+
+        #used to manage names
+        self._next_free_marketplace_number = 0
+
+        #used to add agents
+        self._marketplace_tick_previous_add = 0
+
+        #<<Note many magic numbers below, for simplicity>>
+        #note: KPIs class also has some magic number
+
+        #as ecosystem improves, these parameters may change / improve
+        self._marketplace_percent_toll_to_ocean = 0.002 #magic number
+        self._percent_burn: float = 0.05 #to burning, vs to DAO #magic number
+
+        self._total_OCEAN_minted: float = 0.0
+        self._total_OCEAN_burned: float = 0.0
+        self._total_OCEAN_burned_USD: float = 0.0
+
+        self._speculation_valuation = 5e6 #in USD #magic number
+        self._percent_increase_speculation_valuation_per_s = 0.10 / S_PER_YEAR # ""
+
+        #track certain metrics over time, so that we don't have to load
+        self.kpis = KPIs(self.ss.time_step)
+
+        log.debug("init: end")
+            
+    def takeStep(self) -> None:
+        """This happens once per tick"""
+
+        #update global state values: revenue, valuation
+        self.kpis.takeStep(self)
+
+        #update global state values: other
+        self._speculation_valuation *= (1.0 + self._percent_increase_speculation_valuation_per_s * self.ss.time_step)
+        print(self._speculation_valuation)
+
+     
+    def marketplacePercentTollToOcean(self) -> float:
+        return self._marketplace_percent_toll_to_ocean
+    
+    def percentToBurn(self) -> float:
+        return self._percent_burn
+
+    def percentToOceanDao(self) -> float:
+        return 1.0 - self._percent_burn
+    
+    #==============================================================
+    # def grantTakersSpentAtTick(self) -> float:
+    #     return sum(
+    #         agent.spentAtTick()
+    #         for agent in self.agents.values()
+    #         if isinstance(agent, GrantTakingAgent))
+
+    #==============================================================
+    def OCEANprice(self) -> float:
+        """Estimated price of $OCEAN token, in USD"""
+        price = valuation.OCEANprice(self.overallValuation(),
+                                     self.OCEANsupply())
+        assert price > 0.0
+        return price
+    
+    #==============================================================
+    def overallValuation(self) -> float: #in USD
+        v = self.fundamentalsValuation() + \
+            self.speculationValuation()
+        assert v > 0.0
+        return v
+    
+    def fundamentalsValuation(self) -> float: #in USD
+        return self.kpis.valuationPS(30.0) #based on P/S=30                     #magic number
+    
+    def speculationValuation(self) -> float: #in USD
+        return self._speculation_valuation
+        
+    #==============================================================
+    def OCEANsupply(self) -> float:
+        """Current OCEAN token supply"""
+        return self.initialOCEAN() \
+            + self.totalOCEANminted() \
+            - self.totalOCEANburned()
+        
+    def initialOCEAN(self) -> float:
+        return INIT_OCEAN_SUPPLY
+        
+    def totalOCEANminted(self) -> float:
+        return self._total_OCEAN_minted
+        
+    def totalOCEANburned(self) -> float:
+        return self._total_OCEAN_burned
+        
+    def totalOCEANburnedUSD(self) -> float:
+        return self._total_OCEAN_burned_USD
+    
+    
+def funcOne():
+    return 1.0
+
